Combine load and reload, add listener for hash changes and improve error handling
